+++ .gitignore
@@ -1,0 +1,3 @@
+.vscode
+.idea
+.DS_Store
