+++ core/.gitignore
@@ -15,3 +15,4 @@ yarn-debug.log*
 yarn-error.log*
 /.changelog
 .npm/
+.DS_Store
