+++ core/package.json
@@ -2,7 +2,7 @@
 	"name": "polkadot-portfolio-core",
 	"version": "0.1.1",
 	"description": "Core engine to scrape the portfolio fo an account in the Polkadot ecosystem.",
-	"main": "build/index.js",
+	"main": "build/index.d.ts",
 	"types": "build/index.d.ts",
 
 	"scripts": {
@@ -35,7 +35,13 @@
 	"dependencies": {
 		"@acala-network/types": "4.0.1",
 		"@open-web3/orml-types": "^1.1.3",
-		"@polkadot/api": "^7.1.1",
+		"@polkadot/api": "7.1.1",
+		"@polkadot/rpc-augment": "7.1.1",
+		"@polkadot/rpc-core": "7.1.1",
+		"@polkadot/api-augment": "7.1.1",
+		"@polkadot/types-codec": "7.1.1",
+		"@polkadot/types-create": "7.1.1",
+		"@polkadot/types": "7.1.1",
 		"axios": "^0.26.1",
 		"bn.js": "^5.2.0",
 		"currency-formatter": "^1.5.9",
