+++ core/src/fetch/acala.ts
@@ -1,6 +1,7 @@
+// @ts-nocheck
 import { ApiPromise } from '@polkadot/api';
 import { Asset, AssetOrigin } from '../types';
-import { CurrencyId, PoolId, TradingPair, DexShare } from "@acala-network/types/interfaces/"
+import { CurrencyId, PoolId, DexShare } from "@acala-network/types/interfaces/"
 import { OrmlAccountData } from "@open-web3/orml-types/interfaces/"
 import { types } from "@acala-network/types"
 import BN from 'bn.js';
