+++ core/src/fetch/index.ts
@@ -1,3 +1,4 @@
 export * from './acala';
 export * from './moonbeam'
 export * from './substrate'
+export * from './scrape'
