+++ core/src/fetch/moonbeam.ts
@@ -1,3 +1,4 @@
+// @ts-nocheck
 import { ApiPromise } from "@polkadot/api";
 import BN from "bn.js";
 import { Asset } from "../types";
