+++ ui/package.json
@@ -3,10 +3,9 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "polkadot-portfolio-core": "../core",
     "bn.js": "^5.2.0",
     "currency-formatter": "^1.5.9",
-
+    "@polkadot/api": "7.1.1",
     "@fortawesome/fontawesome-svg-core": "^6.1.1",
     "@fortawesome/free-solid-svg-icons": "^6.1.1",
     "@fortawesome/react-fontawesome": "^0.1.18",
@@ -22,7 +21,8 @@
     "react-dom": "^18.0.0",
     "react-scripts": "5.0.1",
     "typescript": "^4.6.3",
-    "web-vitals": "^2.1.4"
+    "web-vitals": "^2.1.4",
+    "polkadot-portfolio-core": "../core"
   },
   "devDependencies": {
     "@types/currency-formatter": "^1.5.1",
@@ -36,7 +36,8 @@
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "lint": "eslint . --ext .ts --fix"
+    "lint": "eslint . --ext .ts --fix",
+    "link": "link ../core"
   },
   "eslintConfig": {
     "extends": [
