+++ ui/src/store/store.d.ts
@@ -1,10 +1,11 @@
 import { ApiPromise } from "@polkadot/api";
 import { Asset } from "./types/Asset";
+import {IApiRegistry} from "polkadot-portfolio-core"
 
 export interface StoreState {
   accounts: IAccount[],
   networks: string[],
-  apiRegistry: Map<string, ApiRegistry>,
+  apiRegistry: IApiRegistry,
   assets: Asset[],
   visibility: IVisibility,
   loading: LoadingStates,
